changed order of fragments/activities for security checkin
