+++ app/src/main/java/com/exotikosteam/exotikos/fragments/SecurityCheckinFragment.java
@@ -11,6 +11,7 @@ import android.view.ViewGroup;
 
 import com.exotikosteam.exotikos.R;
 import com.exotikosteam.exotikos.databinding.FragmentSecurityCheckinBinding;
+import com.exotikosteam.exotikos.interfaces.OnButtonsClicks;
 
 /**
  * Created by lramaswamy on 11/15/16.
@@ -19,7 +20,7 @@ import com.exotikosteam.exotikos.databinding.FragmentSecurityCheckinBinding;
 public class SecurityCheckinFragment extends Fragment {
 
     FragmentSecurityCheckinBinding securityCheckinBinding;
-    TravelPrepFragment.OnButtonsClicks listener;
+    OnButtonsClicks listener;
 
     public static SecurityCheckinFragment newInstance(boolean fromFragment) {
             SecurityCheckinFragment fragmentSecurity = new SecurityCheckinFragment();
@@ -33,8 +34,8 @@ public class SecurityCheckinFragment extends Fragment {
     @Override
     public void onAttach(Context context) {
         super.onAttach(context);
-        if(context instanceof TravelPrepFragment.OnButtonsClicks) {
-            listener = (TravelPrepFragment.OnButtonsClicks) context;
+        if(context instanceof OnButtonsClicks) {
+            listener = (OnButtonsClicks) context;
         }
     }
 
@@ -42,14 +43,13 @@ public class SecurityCheckinFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         securityCheckinBinding = DataBindingUtil.inflate(inflater, R.layout.fragment_security_checkin, container, false);
+
+        //hide the next button if it is in the activity, instead of the fragment
         boolean fromFragment = getArguments().getBoolean("fromFragment");
         if(!fromFragment)
             securityCheckinBinding.btnNext.setVisibility(View.GONE);
         else
             securityCheckinBinding.btnNext.setVisibility(View.VISIBLE);
-        securityCheckinBinding.btnSecurityVideo.setOnClickListener(v -> {
-            handleLaunchSecurityVidoeHelpPage();
-        });
 
         securityCheckinBinding.btnNext.setOnClickListener(v -> {
             handleLaunchSecurityActivity();
@@ -58,13 +58,7 @@ public class SecurityCheckinFragment extends Fragment {
         return securityCheckinBinding.getRoot();
     }
 
-    private void handleLaunchSecurityVidoeHelpPage() {
-        listener.handleButtonsClicks("LaunchSecurityCheckinVideoHelpPage");
-    }
 
-    private void handleLaunchSecurityHelpPage() {
-        listener.handleButtonsClicks("LaunchSecurityCheckinHelpPage");
-    }
 
     private void handleLaunchSecurityActivity() {
         listener.handleButtonsClicks("LaunchSecurityCheckinActivity");
