+++ app/src/main/java/com/exotikosteam/exotikos/interfaces/OnButtonsClicks.java
@@ -1,0 +1,5 @@
+package com.exotikosteam.exotikos.interfaces;
+
+public interface OnButtonsClicks {
+        void handleButtonsClicks(String buttonName);
+    }
