+++ app/src/main/res/layout/fragment_security_help.xml
@@ -1,6 +1,6 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+<LinearLayout
     android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="wrap_content">
@@ -16,5 +16,11 @@
         android:paddingRight="@dimen/fragment_horizontal_margin"
         android:id="@+id/imageView" />
 
-
+    <Button
+        android:background="@drawable/youtube_play"
+        android:layout_width="200dp"
+        android:layout_height="200dp"
+        android:layout_gravity="center"
+        android:id="@+id/btnSecurityVideo" />
 </LinearLayout>
+</layout>
